Fixes: #346 - docker-entrypoint for zammad-nginx fails to adapt nginx config properly. (#372)
This adds a list of variables that will be forwarded to the containers, but only if they are set via .env file or another mechanism. Otherwise, the containers will not see them.

https://docs.docker.com/compose/compose-file/05-services/#environment

This removes the need to add these variables to docker-compose.override.yml to ensure they get passed to the containers. Now it is enough to specify them via .env.